+++ changelog.md
@@ -5,7 +5,8 @@
 - progress bar support added for ffmpeg audio encoding
 - l-smash parameters prefer_hw and format have menu support in the avs/vs code editor
 - update: l-smash-works 2019-09-10 HolyWu
-- a format exception was fixed
+- a format exception was fixed, it was happening when the cut/trim feature was used
+- the container/muxer dialog has a new tab for MKV attachments
 
 ### 2.0.4.0 Stable